+++ ChangeLog.md
@@ -2,3 +2,9 @@
     updating 0.0.9 versions in Workspace8 2019 for submission
     cleaned up (mostly)
 0.1.2 preparing for 9.2.1
+  
+0.1.3: added stack build lts 19.16 for ghc 9.0.2
+
+0.1.3.3 split pointless and tuples to produce a tuples with nothing else module 
+    aligning with the function names from algebra of programming (import)
+    import tuples from extra (to reduce duplications)